requirements.txt
langchain
openai
faiss-cpu
python-dotenv
gradio
pypdf
tiktoken
